backend/requirements.txt
Django==5.2
django-ckeditor==6.7.2
django-jazzmin==3.0.1
Pillow==11.1.0
requests==2.32.3
